refactor: split chat interface into focused components
