+++ app/src/main/java/com/echoflow/ui/screens/ChatScreen.kt
@@ -1,3 +1,4 @@
+
 @file:OptIn(ExperimentalFoundationApi::class, ExperimentalMaterial3Api::class, ExperimentalMaterial3ExpressiveApi::class, ExperimentalLayoutApi::class)
 
 package com.echoflow.ui.screens
@@ -531,1635 +532,4 @@ fun ChatScreen(
             )
         }
     }
-}
-
-/**
- * The scrolling message list for one conversation. Owns its own [LazyListState] so each chat keeps
- * its scroll position and a switch (via the parent key()) doesn't inherit the previous chat's
- * offset. Keeps the stick-to-bottom behaviour (respects manual scroll, follows streaming).
- */
-@Composable
-private fun MessagesPane(
-    messages: List<ChatMessage>,
-    isStreaming: Boolean,
-    segments: List<StreamSegment>,
-    statusNote: String?,
-    progressLoading: Boolean,
-    modelLoading: Boolean,
-    researchRun: ResearchRun?,
-    onCancelResearch: () -> Unit,
-    topInset: Dp = Spacing.l,
-    bottomInset: Dp = Spacing.l,
-    onCopy: (String) -> Unit,
-    onArtifactOpen: () -> Unit = {},
-) {
-    val listState = rememberLazyListState()
-    var autoFollow by remember { mutableStateOf(true) }
-    val atBottom by remember {
-        derivedStateOf {
-            val li = listState.layoutInfo
-            val last = li.visibleItemsInfo.lastOrNull() ?: return@derivedStateOf true
-            last.index >= li.totalItemsCount - 1 && (last.offset + last.size) <= li.viewportEndOffset + 8
-        }
-    }
-    LaunchedEffect(atBottom, listState.isScrollInProgress) {
-        if (listState.isScrollInProgress) autoFollow = atBottom
-    }
-    LaunchedEffect(messages.size, progressLoading) {
-        if (autoFollow) {
-            val idx = listState.layoutInfo.totalItemsCount - 1
-            if (idx >= 0) runCatching { listState.scrollToItem(idx, Int.MAX_VALUE) }
-        }
-    }
-    LaunchedEffect(autoFollow, isStreaming, progressLoading) {
-        if (autoFollow && (isStreaming || progressLoading)) {
-            while (true) {
-                withFrameNanos { it }
-                if (!listState.isScrollInProgress) {
-                    val idx = listState.layoutInfo.totalItemsCount - 1
-                    if (idx >= 0) runCatching { listState.scrollToItem(idx, Int.MAX_VALUE) }
-                }
-            }
-        }
-    }
-    LazyColumn(
-        state = listState,
-        modifier = Modifier.fillMaxSize(),
-        verticalArrangement = Arrangement.spacedBy(Spacing.l),
-        contentPadding = PaddingValues(start = Spacing.base, end = Spacing.base, top = topInset, bottom = bottomInset),
-    ) {
-        items(messages, key = { it.id }) { msg ->
-            MessageBubble(msg, onCopy = { onCopy(msg.content) }, onArtifactOpen = onArtifactOpen)
-        }
-        researchRun?.let { run ->
-            item(key = "research") { ResearchProgressCard(run = run, onCancel = onCancelResearch) }
-        }
-        if (modelLoading && segments.isEmpty()) {
-            item { ModelLoadingRow() }
-        } else if (progressLoading && segments.isEmpty()) {
-            item { ThinkingRow() }
-        }
-        if (segments.isNotEmpty()) item(key = "streaming") {
-            StreamingAssistantBubble(segments = segments, statusNote = statusNote, isStreaming = isStreaming, onArtifactOpen = onArtifactOpen)
-        }
-    }
-}
-
-/**
- * Shown while an on-device model is loaded into RAM (or a long chat is prefilled) —
- * this can take several seconds to a minute on first use, so it gets a richer
- * animation than the plain "Thinking…" row: breathing brand mark plus an
- * indeterminate expressive wavy progress line.
- */
-@Composable
-private fun ModelLoadingRow(modifier: Modifier = Modifier) {
-    Column(modifier.fillMaxWidth()) {
-        Row(verticalAlignment = Alignment.CenterVertically) {
-            BrandMark(modifier = Modifier.padding(end = Spacing.m), size = 32.dp, animated = true)
-            Column(Modifier.weight(1f)) {
-                Text(
-                    "Loading model…",
-                    style = MaterialTheme.typography.bodyMedium,
-                    color = MaterialTheme.colorScheme.onSurface,
-                )
-                Text(
-                    "Getting it into memory — first use takes the longest",
-                    style = MaterialTheme.typography.bodySmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                )
-            }
-        }
-        Spacer(Modifier.height(Spacing.m))
-        LinearWavyProgressIndicator(
-            modifier = Modifier.fillMaxWidth().padding(end = Spacing.xl),
-            color = MaterialTheme.colorScheme.primary,
-        )
-    }
-}
-
-/**
- * The live assistant reply, rendered as an ordered timeline: reasoning traces, web search
- * steps and text blocks appear in the order the model produced them — so when the model
- * searches, writes, then searches again, the user sees exactly that.
- */
-@Composable
-private fun StreamingAssistantBubble(
-    segments: List<StreamSegment>,
-    statusNote: String?,
-    isStreaming: Boolean,
-    onArtifactOpen: () -> Unit = {},
-) {
-    Column(Modifier.fillMaxWidth()) {
-        Row(verticalAlignment = Alignment.CenterVertically) {
-            BrandMark(size = 26.dp, animated = true)
-            Spacer(Modifier.width(Spacing.s))
-            Text("EchoFlow", style = MaterialTheme.typography.labelLarge, color = MaterialTheme.colorScheme.primary)
-        }
-        Spacer(Modifier.height(Spacing.s))
-
-        segments.forEachIndexed { index, segment ->
-            val isLast = index == segments.lastIndex
-            key(index) {
-                when (segment) {
-                    is StreamSegment.Reasoning -> {
-                        ReasoningSection(reasoning = segment.text, active = isStreaming && isLast)
-                        Spacer(Modifier.height(Spacing.s))
-                    }
-                    is StreamSegment.Search -> {
-                        SearchActivityCard(query = segment.query, sources = segment.sources, active = segment.active)
-                        Spacer(Modifier.height(Spacing.s))
-                    }
-                    is StreamSegment.Advisor -> {
-                        AdvisorCard(
-                            advisorName = segment.advisorName,
-                            advisorModel = segment.advisorModel,
-                            prompt = segment.prompt,
-                            advice = segment.advice,
-                            active = segment.active,
-                        )
-                        Spacer(Modifier.height(Spacing.s))
-                    }
-                    is StreamSegment.Fusion -> {
-                        FusionCard(
-                            panelName = segment.panelName,
-                            models = segment.models,
-                            analysis = segment.analysis,
-                            active = segment.active,
-                        )
-                        Spacer(Modifier.height(Spacing.s))
-                    }
-                    is StreamSegment.AgentRun -> {
-                        AgentDeployingCard()
-                        Spacer(Modifier.height(Spacing.s))
-                    }
-                    is StreamSegment.Subagent -> {
-                        SubagentCard(
-                            taskName = segment.taskName,
-                            taskDescription = segment.taskDescription,
-                            workerModel = segment.workerModel,
-                            outcome = segment.outcome,
-                            error = segment.error,
-                            active = segment.active,
-                        )
-                        Spacer(Modifier.height(Spacing.s))
-                    }
-                    is StreamSegment.Artifact -> {
-                        ArtifactCard(
-                            title = segment.title,
-                            artifactType = segment.artifactType,
-                            version = segment.version,
-                            building = segment.building,
-                            charCount = segment.charCount,
-                            truncated = segment.truncated,
-                            onOpen = onArtifactOpen,
-                        )
-                        Spacer(Modifier.height(Spacing.s))
-                    }
-                    is StreamSegment.Text -> {
-                        SmoothStreamingText(segment.text, Modifier.fillMaxWidth())
-                        if (!isLast) Spacer(Modifier.height(Spacing.s))
-                    }
-                }
-            }
-        }
-
-        statusNote?.let { note ->
-            Spacer(Modifier.height(Spacing.s))
-            Text(
-                note,
-                style = MaterialTheme.typography.labelMedium,
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-            )
-        }
-    }
-}
-
-@Composable
-private fun ChatTopBar(modelName: String, researchMode: Boolean, onMenu: () -> Unit, onModel: () -> Unit, onNewChat: () -> Unit, modifier: Modifier = Modifier) {
-    CenterAlignedTopAppBar(
-        modifier = modifier,
-        navigationIcon = {
-            // Fun shaped icon button (morphs on press), vividly themed.
-            ShapedIconButton(
-                onClick = onMenu, enabled = true, size = 44.dp,
-                restShape = MaterialShapes.Cookie4Sided, pressedShape = MaterialShapes.Cookie7Sided,
-                container = MaterialTheme.colorScheme.primaryContainer,
-            ) { Icon(Icons.Default.Menu, "Open conversations", Modifier.size(20.dp), tint = MaterialTheme.colorScheme.onPrimaryContainer) }
-        },
-        title = {
-            // Model selector as a Material 3 Expressive split button — both halves open the picker.
-            // In Deep Research mode it selects the research engine instead of the chat model.
-            SplitButtonLayout(
-                leadingButton = {
-                    SplitButtonDefaults.LeadingButton(onClick = onModel) {
-                        if (researchMode) {
-                            Icon(Icons.Default.Science, null, Modifier.size(16.dp), tint = MaterialTheme.colorScheme.primary)
-                            Spacer(Modifier.width(Spacing.xs))
-                        }
-                        Text(
-                            modelName,
-                            style = MaterialTheme.typography.titleSmall,
-                            maxLines = 1, overflow = TextOverflow.Ellipsis,
-                            modifier = Modifier.widthIn(max = 150.dp),
-                        )
-                    }
-                },
-                trailingButton = {
-                    // TrailingButton is a toggle in M3; we just use its tap to open the picker.
-                    SplitButtonDefaults.TrailingButton(checked = false, onCheckedChange = { onModel() }) {
-                        Icon(Icons.Default.KeyboardArrowDown, "Choose model", Modifier.size(20.dp))
-                    }
-                },
-            )
-        },
-        actions = {
-            // Fun shaped icon button that pops + morphs on click, vividly themed.
-            ShapedIconButton(
-                onClick = onNewChat, enabled = true, size = 44.dp,
-                restShape = MaterialShapes.Cookie7Sided, pressedShape = MaterialShapes.Sunny,
-                container = MaterialTheme.colorScheme.tertiaryContainer,
-                pulseOnClick = true,
-            ) { Icon(Icons.Default.Add, "New conversation", Modifier.size(20.dp), tint = MaterialTheme.colorScheme.onTertiaryContainer) }
-        },
-        colors = TopAppBarDefaults.centerAlignedTopAppBarColors(containerColor = Color.Transparent),
-    )
-}
-
-@Composable
-private fun ErrorBanner(message: String, onDismiss: () -> Unit) {
-    Surface(
-        color = MaterialTheme.colorScheme.errorContainer,
-        shape = MaterialTheme.shapes.large,
-        modifier = Modifier.fillMaxWidth().padding(horizontal = Spacing.base, vertical = Spacing.s),
-    ) {
-        Row(Modifier.padding(Spacing.base), verticalAlignment = Alignment.CenterVertically) {
-            Icon(Icons.Default.ErrorOutline, null, tint = MaterialTheme.colorScheme.onErrorContainer)
-            Spacer(Modifier.width(Spacing.m))
-            Text(message, color = MaterialTheme.colorScheme.onErrorContainer, style = MaterialTheme.typography.bodyMedium, modifier = Modifier.weight(1f))
-            IconButton(onClick = onDismiss) { Icon(Icons.Default.Close, "Dismiss", tint = MaterialTheme.colorScheme.onErrorContainer) }
-        }
-    }
-}
-
-@Composable
-private fun MessageBubble(message: ChatMessage, modifier: Modifier = Modifier, streaming: Boolean = false, onArtifactOpen: () -> Unit = {}, onCopy: () -> Unit) {
-    val isUser = message.role == "user"
-    if (isUser) {
-        Row(modifier.fillMaxWidth(), horizontalArrangement = Arrangement.End) {
-            Column(horizontalAlignment = Alignment.End, modifier = Modifier.widthIn(max = 320.dp)) {
-                message.localAttachmentUri?.let { uri ->
-                    MessageAttachmentPreview(
-                        uri = uri,
-                        mimeType = message.localAttachmentMimeType,
-                        name = message.localAttachmentName,
-                        modifier = Modifier.padding(bottom = Spacing.s),
-                    )
-                }
-                Surface(
-                    shape = RoundedCornerShape(26.dp, 26.dp, 8.dp, 26.dp),
-                    color = MaterialTheme.colorScheme.primary,
-                    contentColor = MaterialTheme.colorScheme.onPrimary,
-                ) {
-                    Text(message.content, style = MaterialTheme.typography.bodyLarge, modifier = Modifier.padding(horizontal = 18.dp, vertical = Spacing.m))
-                }
-            }
-        }
-    } else {
-        // ChatGPT / Claude style: no bubble, full content width.
-        Column(modifier.fillMaxWidth()) {
-            Row(verticalAlignment = Alignment.CenterVertically) {
-                BrandMark(size = 26.dp, animated = streaming)
-                Spacer(Modifier.width(Spacing.s))
-                Text("EchoFlow", style = MaterialTheme.typography.labelLarge, color = MaterialTheme.colorScheme.primary)
-            }
-            Spacer(Modifier.height(Spacing.s))
-
-            // Finished replies render their persisted timeline in arrival order, so
-            // reason → search → reason → search → answer keeps exactly the layout it
-            // streamed with instead of merging all reasoning into one block.
-            val persistedSegments = remember(message.id) { ToolEventJson.segmentsFromJson(message.segmentsJson) }
-
-            message.localAttachmentUri?.let { uri ->
-                MessageAttachmentPreview(
-                    uri = uri,
-                    mimeType = message.localAttachmentMimeType,
-                    name = message.localAttachmentName,
-                    modifier = Modifier.padding(bottom = Spacing.s),
-                )
-            }
-
-            when {
-                streaming -> {
-                    // Live markdown, revealed at a smooth steady cadence (decoupled from bursty chunks).
-                    if (message.content.isNotBlank()) SmoothStreamingText(message.content, Modifier.fillMaxWidth())
-                }
-                persistedSegments.isNotEmpty() -> {
-                    val planSteps = remember(message.id) {
-                        persistedSegments.firstOrNull { it.type == "plan" }?.text
-                            ?.split("\n")?.map { it.trim() }?.filter { it.isNotEmpty() } ?: emptyList()
-                    }
-                    val reportCitations = remember(message.id) { ToolEventJson.citationsFromJson(message.citationsJson) }
-                    persistedSegments.forEachIndexed { index, segment ->
-                        when (segment.type) {
-                            "reasoning" -> {
-                                ReasoningSection(reasoning = segment.text.orEmpty(), active = false)
-                                Spacer(Modifier.height(Spacing.s))
-                            }
-                            "search" -> {
-                                SearchActivityCard(query = segment.query.orEmpty(), sources = segment.sources.orEmpty(), active = false)
-                                Spacer(Modifier.height(Spacing.s))
-                            }
-                            "advisor" -> {
-                                segment.advisor?.let { a ->
-                                    AdvisorCard(
-                                        advisorName = a.advisorName,
-                                        advisorModel = a.advisorModel,
-                                        prompt = a.prompt,
-                                        advice = a.advice,
-                                        active = false,
-                                    )
-                                    Spacer(Modifier.height(Spacing.s))
-                                }
-                            }
-                            "fusion" -> {
-                                segment.fusion?.let { f ->
-                                    FusionCard(
-                                        panelName = f.panelName,
-                                        models = f.models,
-                                        analysis = f,
-                                        active = false,
-                                    )
-                                    Spacer(Modifier.height(Spacing.s))
-                                }
-                            }
-                            "subagent" -> {
-                                segment.subagent?.let { s ->
-                                    SubagentCard(
-                                        taskName = s.taskName,
-                                        taskDescription = s.taskDescription,
-                                        workerModel = s.workerModel,
-                                        outcome = s.outcome,
-                                        error = s.error,
-                                        active = false,
-                                    )
-                                    Spacer(Modifier.height(Spacing.s))
-                                }
-                            }
-                            "artifact" -> {
-                                segment.artifact?.let { a ->
-                                    ArtifactCard(
-                                        title = a.title,
-                                        artifactType = a.type,
-                                        version = a.version,
-                                        building = false,
-                                        charCount = 0,
-                                        truncated = false,
-                                        onOpen = onArtifactOpen,
-                                    )
-                                    if (index != persistedSegments.lastIndex) Spacer(Modifier.height(Spacing.s))
-                                }
-                            }
-                            "stopped" -> {
-                                StoppedNotice()
-                                if (index != persistedSegments.lastIndex) Spacer(Modifier.height(Spacing.s))
-                            }
-                            // The plan is rendered as a disclosure inside the report card.
-                            "plan" -> Unit
-                            "report" -> {
-                                ReportCard(
-                                    report = segment.text.orEmpty(),
-                                    citations = reportCitations,
-                                    planSteps = planSteps,
-                                    onCopy = onCopy,
-                                )
-                                if (index != persistedSegments.lastIndex) Spacer(Modifier.height(Spacing.s))
-                            }
-                            "data" -> {
-                                DataResultCard(
-                                    json = segment.text.orEmpty(),
-                                    citations = reportCitations,
-                                    onCopy = onCopy,
-                                )
-                                if (index != persistedSegments.lastIndex) Spacer(Modifier.height(Spacing.s))
-                            }
-                            else -> {
-                                RichMarkdown(segment.text.orEmpty(), Modifier.fillMaxWidth())
-                                if (index != persistedSegments.lastIndex) Spacer(Modifier.height(Spacing.s))
-                            }
-                        }
-                    }
-                }
-                else -> {
-                    // Legacy messages saved before the timeline column existed.
-                    val reasoningText = message.reasoning
-                    if (!reasoningText.isNullOrBlank()) {
-                        ReasoningSection(reasoning = reasoningText, active = false)
-                        Spacer(Modifier.height(Spacing.s))
-                    }
-                    val toolEvents = remember(message.id) { ToolEventJson.toolEventsFromJson(message.toolEventsJson) }
-                    toolEvents.forEach { event ->
-                        SearchActivityCard(query = event.query, sources = event.sources, active = false)
-                        Spacer(Modifier.height(Spacing.s))
-                    }
-                    RichMarkdown(message.content, Modifier.fillMaxWidth())
-                }
-            }
-
-            if (!streaming) {
-                Spacer(Modifier.height(Spacing.xs))
-                FilledTonalIconButton(
-                    onClick = onCopy,
-                    modifier = Modifier.size(32.dp),
-                    colors = IconButtonDefaults.filledTonalIconButtonColors(containerColor = MaterialTheme.colorScheme.surfaceContainerHigh),
-                ) { Icon(Icons.Default.ContentCopy, "Copy", Modifier.size(16.dp)) }
-            }
-        }
-    }
-}
-
-/**
- * Smooth "typewriter" reveal, the way production AI apps (T3 Chat, Vercel v0, ChatGPT) do it:
- * the network delivers text in bursts, but we reveal it at a steady, frame-synced cadence so it
- * reads pleasantly instead of flickering in chunks. The pace is a gentle base speed plus a
- * proportional catch-up, so it never lags far behind a fast model yet never feels rushed.
- */
-@Composable
-private fun SmoothStreamingText(
-    text: String,
-    modifier: Modifier = Modifier,
-    markdown: Boolean = true,
-    style: androidx.compose.ui.text.TextStyle = MaterialTheme.typography.bodyLarge,
-    color: Color = MaterialTheme.colorScheme.onSurface,
-) {
-    val target by rememberUpdatedState(text)
-    var shown by remember { mutableStateOf(0) }
-    LaunchedEffect(Unit) {
-        var lastFrame = 0L
-        while (true) {
-            val frame = withFrameNanos { it }
-            val dt = if (lastFrame == 0L) 0f else (frame - lastFrame) / 1_000_000_000f
-            lastFrame = frame
-            val t = target
-            if (shown > t.length) shown = 0 // a new message started — restart the reveal
-            if (shown < t.length) {
-                val remaining = t.length - shown
-                // Steady, pleasant cadence that scales with backlog so the whole answer finishes a
-                // beat after the model (≈2s drain), never instant-dumping even at 1000+ tps.
-                val charsPerSec = (remaining / 2f).coerceIn(40f, 900f)
-                val add = (charsPerSec * dt).toInt().coerceAtLeast(1)
-                shown = (shown + add).coerceAtMost(t.length)
-            }
-        }
-    }
-    val revealed = target.take(shown)
-    if (markdown) {
-        // Live markdown while streaming. Because the text is revealed gradually (not in bursts),
-        // markdown spans complete one char at a time, so re-layout stays smooth like ChatGPT/Claude.
-        MarkdownText(text = revealed, modifier = modifier, textColor = color, style = style)
-    } else {
-        SelectionContainer { Text(text = revealed, style = style, color = color, modifier = modifier) }
-    }
-}
-
-/**
- * Collapsible reasoning ("thinking") trace, styled like T3 Chat / Claude. Subtly tinted so it reads
- * as meta-content. Auto-expands and streams while the model is reasoning, then auto-collapses once
- * the answer begins; the user can expand/collapse at any time (collapsed by default once complete).
- */
-@Composable
-private fun ReasoningSection(reasoning: String, active: Boolean) {
-    var userToggled by remember { mutableStateOf<Boolean?>(null) }
-    val expanded = userToggled ?: active
-    val chevron by animateFloatAsState(if (expanded) 180f else 0f, label = "reasoning-chevron")
-    Surface(
-        onClick = { userToggled = !expanded },
-        shape = MaterialTheme.shapes.large,
-        color = MaterialTheme.colorScheme.tertiaryContainer.copy(alpha = 0.35f),
-        modifier = Modifier.fillMaxWidth(),
-    ) {
-        Column(Modifier.padding(horizontal = Spacing.base, vertical = Spacing.m)) {
-            Row(verticalAlignment = Alignment.CenterVertically) {
-                Icon(Icons.Default.Psychology, null, Modifier.size(18.dp), tint = MaterialTheme.colorScheme.tertiary)
-                Spacer(Modifier.width(Spacing.s))
-                Text(
-                    if (active) "Reasoning…" else "Reasoning",
-                    style = MaterialTheme.typography.labelLarge,
-                    color = MaterialTheme.colorScheme.onSurface,
-                )
-                if (active) {
-                    Spacer(Modifier.width(Spacing.s))
-                    LoadingIndicator(color = MaterialTheme.colorScheme.tertiary, modifier = Modifier.size(18.dp))
-                }
-                Spacer(Modifier.weight(1f))
-                Icon(
-                    Icons.Default.KeyboardArrowDown, if (expanded) "Collapse" else "Expand",
-                    Modifier.size(20.dp).rotate(chevron),
-                    tint = MaterialTheme.colorScheme.onSurfaceVariant,
-                )
-            }
-            AnimatedVisibility(visible = expanded, enter = expandVertically() + fadeIn(), exit = shrinkVertically() + fadeOut()) {
-                if (active) {
-                    // Bounded, internally auto-scrolling panel so a fast reasoning stream can never
-                    // overflow / "break out" of the container — it scrolls within a fixed height.
-                    val sc = rememberScrollState()
-                    LaunchedEffect(sc.maxValue) { sc.scrollTo(sc.maxValue) }
-                    Box(
-                        Modifier
-                            .padding(top = Spacing.s)
-                            .fillMaxWidth()
-                            .heightIn(max = 190.dp)
-                            .verticalScroll(sc),
-                    ) {
-                        SmoothStreamingText(
-                            reasoning, Modifier.fillMaxWidth(),
-                            markdown = true,
-                            style = MaterialTheme.typography.bodyMedium,
-                            color = MaterialTheme.colorScheme.onSurfaceVariant,
-                        )
-                    }
-                } else {
-                    Box(Modifier.padding(top = Spacing.s)) {
-                        MarkdownText(
-                            reasoning, Modifier.fillMaxWidth(),
-                            textColor = MaterialTheme.colorScheme.onSurfaceVariant,
-                            style = MaterialTheme.typography.bodyMedium,
-                        )
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun ThinkingRow(modifier: Modifier = Modifier) {
-    Row(modifier, verticalAlignment = Alignment.CenterVertically) {
-        BrandMark(modifier = Modifier.padding(end = Spacing.m), size = 32.dp, animated = true)
-        LoadingIndicator(color = MaterialTheme.colorScheme.primary)
-        Spacer(Modifier.width(Spacing.s))
-        Text("Thinking…", style = MaterialTheme.typography.bodyMedium, color = MaterialTheme.colorScheme.onSurfaceVariant)
-    }
-}
-
-/** Red "Message stopped" line shown under a reply the user cancelled with the Stop button. */
-@Composable
-private fun StoppedNotice(modifier: Modifier = Modifier) {
-    Row(modifier, verticalAlignment = Alignment.CenterVertically) {
-        Icon(
-            Icons.Default.Stop,
-            null,
-            Modifier.size(16.dp),
-            tint = MaterialTheme.colorScheme.error,
-        )
-        Spacer(Modifier.width(Spacing.xs))
-        Text(
-            "Message stopped",
-            style = MaterialTheme.typography.labelMedium,
-            color = MaterialTheme.colorScheme.error,
-        )
-    }
-}
-
-@Composable
-private fun EmptyState(onSuggestion: (String) -> Unit) {
-    Column(
-        Modifier.fillMaxSize().padding(horizontal = Spacing.xl),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center,
-    ) {
-        // Haloed, morphing hero — a strong colored focal point.
-        Box(contentAlignment = Alignment.Center) {
-            val morph = rememberMorph(BrandShapes.heroStart, BrandShapes.heroEnd)
-            val progress by rememberMorphProgress(3400)
-            Box(
-                Modifier.size(150.dp).clip(MorphPolygonShape(morph, progress)).background(MaterialTheme.colorScheme.primaryContainer),
-            )
-            BrandMark(size = 84.dp, animated = true, iconScale = 0.42f)
-        }
-        Spacer(Modifier.height(Spacing.xl))
-        Text(
-            "How can I help\nyou today?",
-            style = MaterialTheme.typography.displaySmall,
-            color = MaterialTheme.colorScheme.onSurface,
-            textAlign = TextAlign.Center,
-        )
-        Spacer(Modifier.height(Spacing.xl))
-
-        data class Sug(val icon: ImageVector, val label: String, val prompt: String, val container: Color, val onContainer: Color)
-        val cs = MaterialTheme.colorScheme
-        val suggestions = listOf(
-            Sug(Icons.Default.Lightbulb, "Explain", "Explain quantum computing in simple terms", cs.primaryContainer, cs.onPrimaryContainer),
-            Sug(Icons.Default.Edit, "Write", "Write an email asking for a deadline extension", cs.secondaryContainer, cs.onSecondaryContainer),
-            Sug(Icons.Default.Map, "Plan", "Plan a 3-day itinerary for Tokyo", cs.tertiaryContainer, cs.onTertiaryContainer),
-            Sug(Icons.Default.Code, "Code", "Write a Python script to rename files in a folder", cs.surfaceContainerHigh, cs.onSurface),
-        )
-        FlowRow(
-            horizontalArrangement = Arrangement.spacedBy(Spacing.s, Alignment.CenterHorizontally),
-            verticalArrangement = Arrangement.spacedBy(Spacing.s),
-            modifier = Modifier.fillMaxWidth(),
-        ) {
-            suggestions.forEach { s ->
-                AssistPill(s.icon, s.label, s.container, s.onContainer) { onSuggestion(s.prompt) }
-            }
-        }
-    }
-}
-
-@Composable
-private fun AssistPill(icon: ImageVector, label: String, container: Color, onContainer: Color, onClick: () -> Unit) {
-    Surface(onClick = onClick, shape = CircleShape, color = container) {
-        Row(
-            Modifier.padding(start = 14.dp, end = 18.dp, top = Spacing.m, bottom = Spacing.m),
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            Icon(icon, null, Modifier.size(18.dp), tint = onContainer)
-            Spacer(Modifier.width(Spacing.s))
-            Text(label, style = MaterialTheme.typography.labelLarge, color = onContainer)
-        }
-    }
-}
-
-@Composable
-private fun MessageAttachmentPreview(
-    uri: String,
-    mimeType: String?,
-    name: String?,
-    modifier: Modifier = Modifier,
-) {
-    if (mimeType.equals("application/pdf", ignoreCase = true)) {
-        Surface(
-            shape = MaterialTheme.shapes.large,
-            color = MaterialTheme.colorScheme.secondaryContainer,
-            modifier = modifier.widthIn(max = 280.dp),
-        ) {
-            Row(Modifier.padding(Spacing.base), verticalAlignment = Alignment.CenterVertically) {
-                Icon(
-                    Icons.Default.PictureAsPdf,
-                    null,
-                    Modifier.size(28.dp),
-                    tint = MaterialTheme.colorScheme.onSecondaryContainer,
-                )
-                Spacer(Modifier.width(Spacing.s))
-                Text(
-                    name ?: "PDF file",
-                    maxLines = 1,
-                    overflow = TextOverflow.Ellipsis,
-                    style = MaterialTheme.typography.bodyMedium,
-                    color = MaterialTheme.colorScheme.onSecondaryContainer,
-                    modifier = Modifier.weight(1f),
-                )
-            }
-        }
-    } else {
-        AsyncImage(
-            uri,
-            null,
-            modifier.size(200.dp).clip(MaterialTheme.shapes.large),
-            contentScale = ContentScale.Crop,
-        )
-    }
-}
-
-@Composable
-private fun InputToolbar(
-    text: String,
-    onText: (String) -> Unit,
-    pendingUri: String?,
-    pendingMime: String?,
-    pendingName: String?,
-    onClearAttachment: () -> Unit,
-    onAttach: () -> Unit,
-    onAttachPdf: () -> Unit,
-    onReceiveImage: (Uri) -> Unit,
-    imageAttachEnabled: Boolean,
-    pdfAttachEnabled: Boolean,
-    isStreaming: Boolean,
-    deepResearchActive: Boolean,
-    webSearchChipOn: Boolean,
-    dataAgentActive: Boolean,
-    dataAgentAvailable: Boolean,
-    echoAdviserActive: Boolean,
-    echoFusionActive: Boolean,
-    echoAgentActive: Boolean,
-    advisorChipLabel: String?,
-    fusionChipLabel: String?,
-    agentChipLabel: String?,
-    onToggleDeepResearch: () -> Unit,
-    onToggleWebSearch: () -> Unit,
-    onToggleDataAgent: () -> Unit,
-    onToggleEchoAdviser: () -> Unit,
-    onToggleEchoFusion: () -> Unit,
-    onToggleEchoAgent: () -> Unit,
-    echoAdviserAvailable: Boolean,
-    echoFusionAvailable: Boolean,
-    echoAgentAvailable: Boolean,
-    browserFlowActive: Boolean,
-    browserFlowAvailable: Boolean,
-    onToggleBrowserFlow: () -> Unit,
-    artifactActive: Boolean,
-    onToggleArtifact: () -> Unit,
-    browserSession: com.echoflow.data.BrowserSession?,
-    browserSteps: List<com.echoflow.data.BrowserStep>,
-    onBrowserOpen: () -> Unit,
-    onBrowserFinish: () -> Unit,
-    onBrowserStop: () -> Unit,
-    onBrowserPick: (String) -> Unit,
-    onBrowserConfirmDomain: () -> Unit,
-    onBrowserConfirmSend: () -> Unit,
-    onBrowserCancel: () -> Unit,
-    researchInProgress: Boolean,
-    researchEngineLabel: String?,
-    dataAgentLabel: String,
-    showEffortPill: Boolean,
-    exaEffort: String,
-    onSelectEffort: (String) -> Unit,
-    blockedReason: String? = null,
-    onSend: () -> Unit,
-    onStop: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    val imageContentReceiver = remember(imageAttachEnabled, onReceiveImage) {
-        object : ReceiveContentListener {
-            override fun onReceive(transferableContent: TransferableContent): TransferableContent? {
-                if (!imageAttachEnabled || !transferableContent.hasMediaType(MediaType.Image)) {
-                    return transferableContent
-                }
-                return transferableContent.consume { item ->
-                    item.uri?.let { uri ->
-                        onReceiveImage(uri)
-                        true
-                    } == true
-                }
-            }
-        }
-    }
-
-    Column(
-        modifier
-            .fillMaxWidth()
-            .windowInsetsPadding(WindowInsets.ime.union(WindowInsets.navigationBars))
-            .padding(horizontal = Spacing.base, vertical = Spacing.m),
-    ) {
-        // Persistent Browser Flow card — the in-chat anchor/control surface for the live session.
-        browserSession?.let { session ->
-            com.echoflow.ui.components.BrowserSessionCard(
-                session = session,
-                steps = browserSteps,
-                onOpen = onBrowserOpen,
-                onFinish = onBrowserFinish,
-                onStop = onBrowserStop,
-                onPickCandidate = onBrowserPick,
-                onConfirmDomain = onBrowserConfirmDomain,
-                onConfirmSend = onBrowserConfirmSend,
-                onCancel = onBrowserCancel,
-                modifier = Modifier.padding(bottom = Spacing.s),
-            )
-        }
-
-        AnimatedVisibility(visible = pendingUri != null) {
-            pendingUri?.let { uri ->
-                Surface(
-                    shape = MaterialTheme.shapes.large,
-                    color = MaterialTheme.colorScheme.secondaryContainer,
-                    modifier = Modifier.padding(bottom = Spacing.s),
-                ) {
-                    Row(Modifier.padding(Spacing.s), verticalAlignment = Alignment.CenterVertically) {
-                        if (pendingMime.equals("application/pdf", ignoreCase = true)) {
-                            Icon(
-                                Icons.Default.PictureAsPdf,
-                                null,
-                                Modifier.size(40.dp),
-                                tint = MaterialTheme.colorScheme.onSecondaryContainer,
-                            )
-                        } else {
-                            AsyncImage(uri, null, Modifier.size(40.dp).clip(MaterialTheme.shapes.medium), contentScale = ContentScale.Crop)
-                        }
-                        Spacer(Modifier.width(Spacing.m))
-                        Text(pendingName ?: "Attachment", maxLines = 1, overflow = TextOverflow.Ellipsis, style = MaterialTheme.typography.bodyMedium, color = MaterialTheme.colorScheme.onSecondaryContainer, modifier = Modifier.weight(1f))
-                        IconButton(onClick = onClearAttachment, modifier = Modifier.size(28.dp)) { Icon(Icons.Default.Close, "Remove", Modifier.size(18.dp), tint = MaterialTheme.colorScheme.onSecondaryContainer) }
-                    }
-                }
-            }
-        }
-
-        // Active capability chips — always show what's on so behaviour is never hidden.
-        AnimatedVisibility(visible = webSearchChipOn || deepResearchActive || dataAgentActive || echoAdviserActive || echoFusionActive || echoAgentActive || browserFlowActive || artifactActive) {
-            FlowRow(
-                horizontalArrangement = Arrangement.spacedBy(Spacing.s),
-                verticalArrangement = Arrangement.spacedBy(Spacing.s),
-                modifier = Modifier.padding(start = Spacing.s, bottom = Spacing.s),
-            ) {
-                if (webSearchChipOn) {
-                    CapabilityChip(Icons.Default.TravelExplore, "Web search", onRemove = onToggleWebSearch)
-                }
-                if (browserFlowActive) {
-                    CapabilityChip(Icons.Default.Language, "Browser Flow", onRemove = onToggleBrowserFlow)
-                }
-                if (artifactActive) {
-                    CapabilityChip(Icons.Default.AutoAwesome, "Artifact", onRemove = onToggleArtifact)
-                }
-                if (echoAdviserActive) {
-                    CapabilityChip(
-                        Icons.Default.Psychology,
-                        advisorChipLabel?.let { "Adviser · $it" } ?: "Echo Adviser",
-                        onRemove = onToggleEchoAdviser,
-                    )
-                }
-                if (echoFusionActive) {
-                    CapabilityChip(
-                        Icons.Default.AccountTree,
-                        fusionChipLabel?.let { "Fusion · $it" } ?: "Echo Fusion",
-                        onRemove = onToggleEchoFusion,
-                    )
-                }
-                if (echoAgentActive) {
-                    CapabilityChip(
-                        Icons.Default.Hub,
-                        agentChipLabel?.let { "Echo Agent · $it" } ?: "Echo Agents",
-                        onRemove = onToggleEchoAgent,
-                    )
-                }
-                if (deepResearchActive) {
-                    CapabilityChip(
-                        Icons.Default.Science,
-                        researchEngineLabel?.takeIf { it.isNotBlank() && it != "Choose engine" }?.let { "Research · $it" } ?: "Deep Research",
-                        onRemove = onToggleDeepResearch,
-                    )
-                    // Exa Agent's depth/cost dial lives here, not in the engine list.
-                    if (showEffortPill) EffortPill(effort = exaEffort, onSelect = onSelectEffort)
-                }
-                if (dataAgentActive) {
-                    CapabilityChip(
-                        Icons.Default.Science,
-                        dataAgentLabel.takeIf { it.isNotBlank() && it != "Choose agent" }?.let { "Data Agent · $it" } ?: "Data Agent",
-                        onRemove = onToggleDataAgent,
-                    )
-                }
-            }
-        }
-
-        AnimatedVisibility(visible = blockedReason != null) {
-            Text(
-                blockedReason.orEmpty(),
-                style = MaterialTheme.typography.labelMedium,
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-                modifier = Modifier.padding(start = Spacing.base, bottom = Spacing.s),
-            )
-        }
-        Surface(
-            shape = CircleShape,
-            // Tint the composer while the chat is captured by a live browser session.
-            color = if (browserSession != null) MaterialTheme.colorScheme.primaryContainer
-            else MaterialTheme.colorScheme.surfaceContainerHigh,
-            tonalElevation = 3.dp,
-            shadowElevation = 8.dp,
-            modifier = Modifier.fillMaxWidth(),
-        ) {
-            Row(Modifier.padding(Spacing.s), verticalAlignment = Alignment.CenterVertically) {
-                var plusMenuOpen by remember { mutableStateOf(false) }
-                Box {
-                    ShapedIconButton(
-                        onClick = { plusMenuOpen = true },
-                        enabled = true,
-                        size = 44.dp,
-                        restShape = MaterialShapes.Cookie6Sided,
-                        pressedShape = MaterialShapes.Flower,
-                        container = MaterialTheme.colorScheme.tertiaryContainer,
-                        pulseOnClick = true,
-                    ) {
-                        Icon(Icons.Default.Add, "Add context or capability", Modifier.size(22.dp), tint = MaterialTheme.colorScheme.onTertiaryContainer)
-                    }
-                    PlusMenu(
-                        expanded = plusMenuOpen,
-                        onDismiss = { plusMenuOpen = false },
-                        showImage = imageAttachEnabled,
-                        showFiles = pdfAttachEnabled,
-                        webSearchOn = webSearchChipOn,
-                        deepResearchOn = deepResearchActive,
-                        dataAgentOn = dataAgentActive,
-                        dataAgentAvailable = dataAgentAvailable,
-                        echoAdviserOn = echoAdviserActive,
-                        echoFusionOn = echoFusionActive,
-                        echoAgentOn = echoAgentActive,
-                        echoAdviserAvailable = echoAdviserAvailable,
-                        echoFusionAvailable = echoFusionAvailable,
-                        echoAgentAvailable = echoAgentAvailable,
-                        browserFlowOn = browserFlowActive,
-                        browserFlowAvailable = browserFlowAvailable,
-                        artifactOn = artifactActive,
-                        onImage = { plusMenuOpen = false; onAttach() },
-                        onFiles = { plusMenuOpen = false; onAttachPdf() },
-                        onToggleWebSearch = { plusMenuOpen = false; onToggleWebSearch() },
-                        onToggleDeepResearch = { plusMenuOpen = false; onToggleDeepResearch() },
-                        onToggleDataAgent = { plusMenuOpen = false; onToggleDataAgent() },
-                        onToggleEchoAdviser = { plusMenuOpen = false; onToggleEchoAdviser() },
-                        onToggleEchoFusion = { plusMenuOpen = false; onToggleEchoFusion() },
-                        onToggleEchoAgent = { plusMenuOpen = false; onToggleEchoAgent() },
-                        onToggleBrowserFlow = { plusMenuOpen = false; onToggleBrowserFlow() },
-                        onToggleArtifact = { plusMenuOpen = false; onToggleArtifact() },
-                    )
-                }
-
-                TextField(
-                    value = text,
-                    onValueChange = onText,
-                    placeholder = {
-                        Text(
-                            when {
-                                browserSession != null -> "Command the browser…"
-                                browserFlowActive -> "Open a site & say what to do…"
-                                dataAgentActive -> "Describe the data to extract…"
-                                deepResearchActive -> "Research a topic…"
-                                artifactActive -> "Describe an artifact to build…"
-                                else -> "Ask anything…"
-                            }
-                        )
-                    },
-                    maxLines = 6,
-                    colors = TextFieldDefaults.colors(
-                        focusedContainerColor = Color.Transparent,
-                        unfocusedContainerColor = Color.Transparent,
-                        disabledContainerColor = Color.Transparent,
-                        focusedIndicatorColor = Color.Transparent,
-                        unfocusedIndicatorColor = Color.Transparent,
-                        disabledIndicatorColor = Color.Transparent,
-                    ),
-                    textStyle = MaterialTheme.typography.bodyLarge,
-                    modifier = Modifier
-                        .weight(1f)
-                        .contentReceiver(imageContentReceiver)
-                        .testTag("chat_input_field"),
-                )
-
-                val researchMode = deepResearchActive || dataAgentActive
-                val hasText = text.trim().isNotEmpty()
-                val hasContent = hasText || (pendingUri != null && !researchMode)
-                val canSend = hasContent && !isStreaming && !researchInProgress && blockedReason == null
-                SendButton(enabled = canSend, isStreaming = isStreaming, research = researchMode, onStop = onStop) {
-                    if (canSend) onSend()
-                }
-            }
-        }
-    }
-}
-
-/** The unified "+" menu: add context (image) or turn on a capability (search / research). */
-@Composable
-private fun PlusMenu(
-    expanded: Boolean,
-    onDismiss: () -> Unit,
-    showImage: Boolean,
-    showFiles: Boolean,
-    webSearchOn: Boolean,
-    deepResearchOn: Boolean,
-    dataAgentOn: Boolean,
-    dataAgentAvailable: Boolean,
-    echoAdviserOn: Boolean,
-    echoFusionOn: Boolean,
-    echoAgentOn: Boolean,
-    echoAdviserAvailable: Boolean,
-    echoFusionAvailable: Boolean,
-    echoAgentAvailable: Boolean,
-    browserFlowOn: Boolean,
-    browserFlowAvailable: Boolean,
-    artifactOn: Boolean,
-    onImage: () -> Unit,
-    onFiles: () -> Unit,
-    onToggleWebSearch: () -> Unit,
-    onToggleDeepResearch: () -> Unit,
-    onToggleDataAgent: () -> Unit,
-    onToggleEchoAdviser: () -> Unit,
-    onToggleEchoFusion: () -> Unit,
-    onToggleEchoAgent: () -> Unit,
-    onToggleBrowserFlow: () -> Unit,
-    onToggleArtifact: () -> Unit,
-) {
-    DropdownMenu(expanded = expanded, onDismissRequest = onDismiss) {
-        // Image is offered for cloud models and vision-capable on-device (.litertlm) bundles.
-        if (showImage) {
-            DropdownMenuItem(
-                text = { Text("Image") },
-                leadingIcon = { Icon(Icons.Outlined.AddPhotoAlternate, null) },
-                onClick = onImage,
-            )
-        }
-        if (showFiles) {
-            DropdownMenuItem(
-                text = { Text("Files") },
-                leadingIcon = { Icon(Icons.Default.PictureAsPdf, null) },
-                onClick = onFiles,
-            )
-        }
-        DropdownMenuItem(
-            text = { Text("Web search") },
-            leadingIcon = { Icon(Icons.Default.TravelExplore, null) },
-            trailingIcon = { if (webSearchOn) Icon(Icons.Default.Check, "On", tint = MaterialTheme.colorScheme.primary) },
-            onClick = onToggleWebSearch,
-        )
-        DropdownMenuItem(
-            text = { Text("Deep Research") },
-            leadingIcon = { Icon(Icons.Default.Science, null) },
-            trailingIcon = { if (deepResearchOn) Icon(Icons.Default.Check, "On", tint = MaterialTheme.colorScheme.primary) },
-            onClick = onToggleDeepResearch,
-        )
-        // Artifacts — build a rendered web page, document or report (model picks the type).
-        DropdownMenuItem(
-            text = { Text("Artifact") },
-            leadingIcon = { Icon(Icons.Default.AutoAwesome, null) },
-            trailingIcon = { if (artifactOn) Icon(Icons.Default.Check, "On", tint = MaterialTheme.colorScheme.primary) },
-            onClick = onToggleArtifact,
-        )
-        // Data Agent only appears once it's enabled in Settings and a Firecrawl key exists.
-        if (dataAgentAvailable) {
-            DropdownMenuItem(
-                text = { Text("Data Agent") },
-                leadingIcon = { Icon(Icons.Default.Dataset, null) },
-                trailingIcon = { if (dataAgentOn) Icon(Icons.Default.Check, "On", tint = MaterialTheme.colorScheme.primary) },
-                onClick = onToggleDataAgent,
-            )
-        }
-        // Browser Flow — a live, stateful browser controlled through chat (needs a Firecrawl key).
-        if (browserFlowAvailable) {
-            DropdownMenuItem(
-                text = { Text("Browser Flow") },
-                leadingIcon = { Icon(Icons.Default.Language, null) },
-                trailingIcon = { if (browserFlowOn) Icon(Icons.Default.Check, "On", tint = MaterialTheme.colorScheme.primary) },
-                onClick = onToggleBrowserFlow,
-            )
-        }
-        // Echo Labs modes — only shown when enabled in Settings → Echo Labs. If the key or a
-        // profile/panel is missing, sending surfaces a "set it up" message.
-        if (echoAdviserAvailable || echoFusionAvailable || echoAgentAvailable) {
-            HorizontalDivider(Modifier.padding(vertical = Spacing.xs))
-        }
-        if (echoAdviserAvailable) {
-            DropdownMenuItem(
-                text = { Text("Echo Adviser") },
-                leadingIcon = { Icon(Icons.Default.Psychology, null) },
-                trailingIcon = { if (echoAdviserOn) Icon(Icons.Default.Check, "On", tint = MaterialTheme.colorScheme.primary) },
-                onClick = onToggleEchoAdviser,
-            )
-        }
-        if (echoFusionAvailable) {
-            DropdownMenuItem(
-                text = { Text("Echo Fusion") },
-                leadingIcon = { Icon(Icons.Default.AccountTree, null) },
-                trailingIcon = { if (echoFusionOn) Icon(Icons.Default.Check, "On", tint = MaterialTheme.colorScheme.primary) },
-                onClick = onToggleEchoFusion,
-            )
-        }
-        if (echoAgentAvailable) {
-            DropdownMenuItem(
-                text = { Text("Echo Agents") },
-                leadingIcon = { Icon(Icons.Default.Hub, null) },
-                trailingIcon = { if (echoAgentOn) Icon(Icons.Default.Check, "On", tint = MaterialTheme.colorScheme.primary) },
-                onClick = onToggleEchoAgent,
-            )
-        }
-    }
-}
-
-@Composable
-private fun SendButton(enabled: Boolean, isStreaming: Boolean, research: Boolean = false, onStop: () -> Unit = {}, onClick: () -> Unit) {
-    if (isStreaming) {
-        // Tappable Stop that stays in visual sync with the "Thinking…" row: the same expressive
-        // LoadingIndicator keeps spinning so the reply still feels live, with a Stop glyph centered
-        // on top so it reads clearly as "tap to stop". Tapping cancels the in-flight reply (cloud
-        // stream or on-device generation).
-        Box(
-            Modifier
-                .size(48.dp)
-                .clip(CircleShape)
-                .background(MaterialTheme.colorScheme.surfaceContainerHighest)
-                .clickable(onClick = onStop),
-            contentAlignment = Alignment.Center,
-        ) {
-            LoadingIndicator(
-                color = MaterialTheme.colorScheme.primary,
-                modifier = Modifier.size(40.dp),
-            )
-            Icon(
-                Icons.Default.Stop,
-                "Stop generating",
-                Modifier.size(18.dp),
-                tint = MaterialTheme.colorScheme.primary,
-            )
-        }
-    } else {
-        // The hero action gets the boldest shape — a "Sunny" that morphs to a rounder cookie on
-        // press with an expressive (bouncy) spring. In research mode it starts the investigation.
-        ShapedIconButton(
-            onClick = onClick,
-            enabled = enabled,
-            size = 48.dp,
-            restShape = MaterialShapes.Sunny,
-            pressedShape = MaterialShapes.Cookie12Sided,
-            container = if (enabled) MaterialTheme.colorScheme.primary else MaterialTheme.colorScheme.surfaceContainerHighest,
-        ) {
-            Icon(
-                if (research) Icons.Default.Science else Icons.AutoMirrored.Filled.Send,
-                if (research) "Start research" else "Send",
-                Modifier.size(20.dp),
-                tint = if (enabled) MaterialTheme.colorScheme.onPrimary else MaterialTheme.colorScheme.onSurfaceVariant,
-            )
-        }
-    }
-}
-
-/**
- * A "fun shape" icon button: filled with a [MaterialShapes] polygon that **morphs to a second
- * shape on press** via the expressive bouncy spring (motion physics). A soft top-lit vertical
- * gradient gives the flat shape a 2.5D sense of volume.
- */
-@Composable
-private fun ShapedIconButton(
-    onClick: () -> Unit,
-    enabled: Boolean,
-    size: Dp,
-    restShape: RoundedPolygon,
-    pressedShape: RoundedPolygon,
-    container: Color,
-    pulseOnClick: Boolean = false,
-    content: @Composable () -> Unit,
-) {
-    val interaction = remember { MutableInteractionSource() }
-    val pressed by interaction.collectIsPressedAsState()
-    val pressProgress by animateFloatAsState(
-        targetValue = if (pressed) 1f else 0f,
-        animationSpec = spring(dampingRatio = Spring.DampingRatioLowBouncy, stiffness = Spring.StiffnessMediumLow),
-        label = "icon-shape-morph",
-    )
-    // One-shot "pop + morph" pulse fired on click (for buttons whose action doesn't change the
-    // screen, like attaching a photo, so the feedback is actually seen).
-    val scope = rememberCoroutineScope()
-    val clickPulse = remember { Animatable(0f) }
-    val progress = maxOf(pressProgress, clickPulse.value)
-    val morph = rememberMorph(restShape, pressedShape)
-    val shape = MorphPolygonShape(morph, progress)
-    val popScale = 1f + 0.18f * clickPulse.value
-    // 2.5D volume: lighter at the top, base colour at the bottom.
-    val brush = Brush.verticalGradient(listOf(lerp(container, Color.White, 0.16f), container))
-    Box(
-        Modifier
-            .size(size)
-            .scale(popScale)
-            .clip(shape)
-            .background(brush)
-            .clickable(
-                interactionSource = interaction,
-                indication = ripple(bounded = true),
-                enabled = enabled,
-                onClick = {
-                    if (pulseOnClick) scope.launch {
-                        clickPulse.snapTo(0f)
-                        clickPulse.animateTo(1f, spring(dampingRatio = Spring.DampingRatioMediumBouncy, stiffness = Spring.StiffnessMedium))
-                        clickPulse.animateTo(0f, tween(durationMillis = 180))
-                    }
-                    onClick()
-                },
-            ),
-        contentAlignment = Alignment.Center,
-    ) { content() }
-}
-
-@Composable
-private fun ModelPickerSheet(
-    models: List<Pair<String, String>>,
-    localModels: List<Pair<String, String>>,
-    selectedId: String,
-    onSelect: (String) -> Unit,
-    onManage: () -> Unit,
-    onDismiss: () -> Unit,
-) {
-    val sheetState = rememberModalBottomSheetState(skipPartiallyExpanded = true)
-    var query by remember { mutableStateOf("") }
-    ModalBottomSheet(
-        onDismissRequest = onDismiss,
-        sheetState = sheetState,
-        containerColor = MaterialTheme.colorScheme.surfaceContainerLow,
-    ) {
-        Column(Modifier.fillMaxWidth().padding(horizontal = Spacing.l)) {
-            Row(verticalAlignment = Alignment.CenterVertically, modifier = Modifier.fillMaxWidth().padding(bottom = Spacing.base)) {
-                Text("Choose a model", style = MaterialTheme.typography.headlineSmall, color = MaterialTheme.colorScheme.onSurface, modifier = Modifier.weight(1f))
-                TextButton(onClick = onManage) {
-                    Icon(Icons.Default.Tune, null, Modifier.size(18.dp))
-                    Spacer(Modifier.width(Spacing.s))
-                    Text("Manage")
-                }
-            }
-            OutlinedTextField(
-                value = query,
-                onValueChange = { query = it },
-                placeholder = { Text("Search models") },
-                leadingIcon = { Icon(Icons.Default.Search, null) },
-                singleLine = true,
-                shape = CircleShape,
-                modifier = Modifier.fillMaxWidth(),
-            )
-            Spacer(Modifier.height(Spacing.m))
-            val filtered = models.filter { it.second.contains(query, true) || it.first.contains(query, true) }
-            val filteredLocal = localModels.filter { it.second.contains(query, true) || it.first.contains(query, true) }
-            if (filtered.isEmpty() && filteredLocal.isEmpty()) {
-                Text(
-                    "No models. Tap Manage to add one in Settings.",
-                    style = MaterialTheme.typography.bodyMedium,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                    modifier = Modifier.padding(vertical = Spacing.xl),
-                )
-            }
-            LazyColumn(
-                Modifier.fillMaxWidth(),
-                verticalArrangement = Arrangement.spacedBy(Spacing.s),
-                contentPadding = PaddingValues(bottom = Spacing.xl),
-            ) {
-                items(filtered, key = { it.first }) { (id, name) ->
-                    ModelRow(name, id, id == selectedId, isLocal = false) { onSelect(id) }
-                }
-                if (filteredLocal.isNotEmpty()) {
-                    item(key = "local-section") {
-                        Box(Modifier.padding(top = Spacing.m)) { SectionLabel("Local & network") }
-                    }
-                    items(filteredLocal, key = { it.first }) { (id, name) ->
-                        ModelRow(name, id, id == selectedId, isLocal = true) { onSelect(id) }
-                    }
-                }
-            }
-        }
-    }
-}
-
-/**
- * Echo Fusion panel picker: each saved panel (a named roster of 2–8 models) is a selectable
- * card showing its members. The chosen panel deliberates on every message until turned off.
- */
-@Composable
-private fun FusionPickerSheet(
-    panels: List<FusionPanel>,
-    selectedId: String,
-    onSelect: (String) -> Unit,
-    onManage: () -> Unit,
-    onDismiss: () -> Unit,
-) {
-    val sheetState = rememberModalBottomSheetState(skipPartiallyExpanded = true)
-    ModalBottomSheet(onDismissRequest = onDismiss, sheetState = sheetState, containerColor = MaterialTheme.colorScheme.surfaceContainerLow) {
-        Column(Modifier.fillMaxWidth().padding(horizontal = Spacing.l)) {
-            Row(verticalAlignment = Alignment.CenterVertically, modifier = Modifier.fillMaxWidth().padding(bottom = Spacing.base)) {
-                Column(Modifier.weight(1f)) {
-                    Text("Choose a fusion panel", style = MaterialTheme.typography.headlineSmall, color = MaterialTheme.colorScheme.onSurface)
-                    Text("Models answer in parallel, a judge compares them", style = MaterialTheme.typography.bodySmall, color = MaterialTheme.colorScheme.onSurfaceVariant)
-                }
-                TextButton(onClick = onManage) {
-                    Icon(Icons.Default.Tune, null, Modifier.size(18.dp)); Spacer(Modifier.width(Spacing.s)); Text("Manage")
-                }
-            }
-            if (panels.isEmpty()) {
-                Column(Modifier.fillMaxWidth().padding(vertical = Spacing.xl), horizontalAlignment = Alignment.CenterHorizontally) {
-                    Icon(Icons.Default.AccountTree, null, Modifier.size(28.dp), tint = MaterialTheme.colorScheme.onSurfaceVariant)
-                    Spacer(Modifier.height(Spacing.s))
-                    Text("No panels yet.\nTap Manage to build one in Settings.", style = MaterialTheme.typography.bodyMedium, color = MaterialTheme.colorScheme.onSurfaceVariant, textAlign = TextAlign.Center)
-                }
-            }
-            LazyColumn(Modifier.fillMaxWidth(), verticalArrangement = Arrangement.spacedBy(Spacing.s), contentPadding = PaddingValues(bottom = Spacing.xl)) {
-                items(panels, key = { it.id }) { panel ->
-                    DrEngineRow(
-                        name = panel.name,
-                        description = panel.names.joinToString(" · ").ifBlank { "${panel.models.size} models" },
-                        selected = panel.id == selectedId,
-                    ) { onSelect(panel.id) }
-                }
-            }
-        }
-    }
-}
-
-/**
- * Echo Adviser picker: two zones — the answering model (any cloud model) and which advisor
- * profile it escalates to. Both persist immediately; the sheet stays open so the user can set
- * both before dismissing.
- */
-@Composable
-private fun AdvisorPickerSheet(
-    models: List<Pair<String, String>>,
-    selectedModelId: String,
-    profiles: List<AdvisorProfile>,
-    selectedProfileId: String,
-    onSelectModel: (String) -> Unit,
-    onSelectProfile: (String) -> Unit,
-    onManage: () -> Unit,
-    onDismiss: () -> Unit,
-) {
-    val sheetState = rememberModalBottomSheetState(skipPartiallyExpanded = true)
-    ModalBottomSheet(onDismissRequest = onDismiss, sheetState = sheetState, containerColor = MaterialTheme.colorScheme.surfaceContainerLow) {
-        Column(Modifier.fillMaxWidth().padding(horizontal = Spacing.l)) {
-            Row(verticalAlignment = Alignment.CenterVertically, modifier = Modifier.fillMaxWidth().padding(bottom = Spacing.base)) {
-                Column(Modifier.weight(1f)) {
-                    Text("Adviser setup", style = MaterialTheme.typography.headlineSmall, color = MaterialTheme.colorScheme.onSurface)
-                    Text("A cloud model answers and consults your advisor", style = MaterialTheme.typography.bodySmall, color = MaterialTheme.colorScheme.onSurfaceVariant)
-                }
-                TextButton(onClick = onManage) {
-                    Icon(Icons.Default.Tune, null, Modifier.size(18.dp)); Spacer(Modifier.width(Spacing.s)); Text("Manage")
-                }
-            }
-
-            Box(Modifier.padding(bottom = Spacing.s)) { SectionLabel("Advisor") }
-            if (profiles.isEmpty()) {
-                Text(
-                    "No advisors yet — tap Manage to set one up.",
-                    style = MaterialTheme.typography.bodySmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                    modifier = Modifier.padding(bottom = Spacing.s),
-                )
-            } else {
-                FlowRow(horizontalArrangement = Arrangement.spacedBy(Spacing.s), verticalArrangement = Arrangement.spacedBy(Spacing.s), modifier = Modifier.padding(bottom = Spacing.m)) {
-                    profiles.forEach { profile ->
-                        FilterChip(
-                            selected = profile.id == selectedProfileId,
-                            onClick = { onSelectProfile(profile.id) },
-                            label = { Text(profile.name) },
-                            leadingIcon = { Icon(Icons.Default.Psychology, null, Modifier.size(16.dp)) },
-                        )
-                    }
-                }
-            }
-
-            Box(Modifier.padding(bottom = Spacing.s)) { SectionLabel("Answering model") }
-            LazyColumn(Modifier.fillMaxWidth().heightIn(max = 360.dp), verticalArrangement = Arrangement.spacedBy(Spacing.s), contentPadding = PaddingValues(bottom = Spacing.m)) {
-                items(models, key = { it.first }) { (id, name) ->
-                    ModelRow(name, id, id == selectedModelId, isLocal = false) { onSelectModel(id) }
-                }
-            }
-
-            Button(onClick = onDismiss, shape = CircleShape, modifier = Modifier.fillMaxWidth().padding(bottom = Spacing.xl)) {
-                Text("Done")
-            }
-        }
-    }
-}
-
-/**
- * Echo Agent picker: two zones — the orchestrator (any cloud model, which drives the toolbox)
- * and which agent profile (the worker model it delegates to). Both persist immediately; the
- * sheet stays open so the user can set both before dismissing.
- */
-@Composable
-private fun AgentPickerSheet(
-    models: List<Pair<String, String>>,
-    selectedModelId: String,
-    profiles: List<AgentProfile>,
-    selectedProfileId: String,
-    onSelectModel: (String) -> Unit,
-    onSelectProfile: (String) -> Unit,
-    onManage: () -> Unit,
-    onDismiss: () -> Unit,
-) {
-    val sheetState = rememberModalBottomSheetState(skipPartiallyExpanded = true)
-    ModalBottomSheet(onDismissRequest = onDismiss, sheetState = sheetState, containerColor = MaterialTheme.colorScheme.surfaceContainerLow) {
-        Column(Modifier.fillMaxWidth().padding(horizontal = Spacing.l)) {
-            Row(verticalAlignment = Alignment.CenterVertically, modifier = Modifier.fillMaxWidth().padding(bottom = Spacing.base)) {
-                Column(Modifier.weight(1f)) {
-                    Text("Echo Agents", style = MaterialTheme.typography.headlineSmall, color = MaterialTheme.colorScheme.onSurface)
-                    Text("Your main model uses tools and hands tasks to your Echo Agent", style = MaterialTheme.typography.bodySmall, color = MaterialTheme.colorScheme.onSurfaceVariant)
-                }
-                TextButton(onClick = onManage) {
-                    Icon(Icons.Default.Tune, null, Modifier.size(18.dp)); Spacer(Modifier.width(Spacing.s)); Text("Manage")
-                }
-            }
-
-            Box(Modifier.padding(bottom = Spacing.s)) { SectionLabel("Echo Agent") }
-            if (profiles.isEmpty()) {
-                Text(
-                    "No Echo Agents yet — tap Manage to set one up.",
-                    style = MaterialTheme.typography.bodySmall,
-                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                    modifier = Modifier.padding(bottom = Spacing.s),
-                )
-            } else {
-                FlowRow(horizontalArrangement = Arrangement.spacedBy(Spacing.s), verticalArrangement = Arrangement.spacedBy(Spacing.s), modifier = Modifier.padding(bottom = Spacing.m)) {
-                    profiles.forEach { profile ->
-                        FilterChip(
-                            selected = profile.id == selectedProfileId,
-                            onClick = { onSelectProfile(profile.id) },
-                            label = { Text(profile.name) },
-                            leadingIcon = { Icon(Icons.Default.Hub, null, Modifier.size(16.dp)) },
-                        )
-                    }
-                }
-            }
-
-            Box(Modifier.padding(bottom = Spacing.s)) { SectionLabel("Main model") }
-            LazyColumn(Modifier.fillMaxWidth().heightIn(max = 360.dp), verticalArrangement = Arrangement.spacedBy(Spacing.s), contentPadding = PaddingValues(bottom = Spacing.m)) {
-                items(models, key = { it.first }) { (id, name) ->
-                    ModelRow(name, id, id == selectedModelId, isLocal = false) { onSelectModel(id) }
-                }
-            }
-
-            Button(onClick = onDismiss, shape = CircleShape, modifier = Modifier.fillMaxWidth().padding(bottom = Spacing.xl)) {
-                Text("Done")
-            }
-        }
-    }
-}
-
-/**
- * Engine picker shown in Deep Research mode. Lists built-in provider-native engines (only
- * those whose API key is configured) plus the user's added agentic chat models. Selecting
- * one stores it as the active Deep Research engine.
- */
-@Composable
-private fun DeepResearchModelSheet(
-    providerEngines: List<DrEngine>,
-    agentModels: List<DeepResearchModel>,
-    selectedId: String,
-    onSelect: (String) -> Unit,
-    onManage: () -> Unit,
-    onDismiss: () -> Unit,
-    title: String = "Deep Research engine",
-    subtitle: String = "Providers run research themselves; chat models orchestrate it",
-    showAgentSection: Boolean = true,
-) {
-    val sheetState = rememberModalBottomSheetState(skipPartiallyExpanded = true)
-    ModalBottomSheet(
-        onDismissRequest = onDismiss,
-        sheetState = sheetState,
-        containerColor = MaterialTheme.colorScheme.surfaceContainerLow,
-    ) {
-        Column(Modifier.fillMaxWidth().padding(horizontal = Spacing.l)) {
-            Row(verticalAlignment = Alignment.CenterVertically, modifier = Modifier.fillMaxWidth().padding(bottom = Spacing.base)) {
-                Column(Modifier.weight(1f)) {
-                    Text(title, style = MaterialTheme.typography.headlineSmall, color = MaterialTheme.colorScheme.onSurface)
-                    Text(
-                        subtitle,
-                        style = MaterialTheme.typography.bodySmall,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant,
-                    )
-                }
-                TextButton(onClick = onManage) {
-                    Icon(Icons.Default.Tune, null, Modifier.size(18.dp))
-                    Spacer(Modifier.width(Spacing.s))
-                    Text("Manage")
-                }
-            }
-
-            if (providerEngines.isEmpty() && (!showAgentSection || agentModels.isEmpty())) {
-                Column(Modifier.fillMaxWidth().padding(vertical = Spacing.xl), horizontalAlignment = Alignment.CenterHorizontally) {
-                    Icon(Icons.Default.Science, null, Modifier.size(28.dp), tint = MaterialTheme.colorScheme.onSurfaceVariant)
-                    Spacer(Modifier.height(Spacing.s))
-                    Text(
-                        "Nothing set up yet.\nAdd a provider key (or model) in Settings.",
-                        style = MaterialTheme.typography.bodyMedium,
-                        color = MaterialTheme.colorScheme.onSurfaceVariant,
-                        textAlign = TextAlign.Center,
-                    )
-                }
-            }
-
-            LazyColumn(
-                Modifier.fillMaxWidth(),
-                verticalArrangement = Arrangement.spacedBy(Spacing.s),
-                contentPadding = PaddingValues(bottom = Spacing.xl),
-            ) {
-                if (providerEngines.isNotEmpty()) {
-                    item(key = "provider-section") { Box(Modifier.padding(top = Spacing.s)) { SectionLabel(if (showAgentSection) "Provider research" else "Agents") } }
-                    items(providerEngines, key = { it.id }) { engine ->
-                        DrEngineRow(engine.name, engine.description, engine.id == selectedId) { onSelect(engine.id) }
-                    }
-                }
-                if (showAgentSection && agentModels.isNotEmpty()) {
-                    item(key = "agent-section") { Box(Modifier.padding(top = Spacing.m)) { SectionLabel("Your research models") } }
-                    items(agentModels, key = { it.id }) { model ->
-                        DrEngineRow(model.name, "Orchestrates searches into a report", model.id == selectedId) { onSelect(model.id) }
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun DrEngineRow(name: String, description: String, selected: Boolean, onClick: () -> Unit) {
-    Surface(
-        onClick = onClick,
-        shape = MaterialTheme.shapes.large,
-        color = if (selected) MaterialTheme.colorScheme.primaryContainer else MaterialTheme.colorScheme.surfaceContainerHigh,
-        modifier = Modifier.fillMaxWidth(),
-    ) {
-        Row(Modifier.padding(Spacing.base), verticalAlignment = Alignment.CenterVertically) {
-            Box(
-                Modifier.size(40.dp).clip(CircleShape).background(
-                    if (selected) MaterialTheme.colorScheme.primary else MaterialTheme.colorScheme.surfaceContainerHighest
-                ),
-                contentAlignment = Alignment.Center,
-            ) {
-                Icon(
-                    Icons.Default.Science, null, Modifier.size(20.dp),
-                    tint = if (selected) MaterialTheme.colorScheme.onPrimary else MaterialTheme.colorScheme.onSurfaceVariant,
-                )
-            }
-            Spacer(Modifier.width(Spacing.base))
-            Column(Modifier.weight(1f)) {
-                Text(
-                    name,
-                    style = MaterialTheme.typography.titleSmall.copy(fontWeight = FontWeight.Bold),
-                    color = if (selected) MaterialTheme.colorScheme.onPrimaryContainer else MaterialTheme.colorScheme.onSurface,
-                    maxLines = 1, overflow = TextOverflow.Ellipsis,
-                )
-                Text(
-                    description,
-                    style = MaterialTheme.typography.bodySmall,
-                    color = if (selected) MaterialTheme.colorScheme.onPrimaryContainer else MaterialTheme.colorScheme.onSurfaceVariant,
-                    maxLines = 1, overflow = TextOverflow.Ellipsis,
-                )
-            }
-            if (selected) Icon(Icons.Default.CheckCircle, null, tint = MaterialTheme.colorScheme.onPrimaryContainer)
-        }
-    }
-}
-
-@Composable
-private fun ModelRow(name: String, modelId: String, selected: Boolean, isLocal: Boolean = false, onClick: () -> Unit) {
-    val displayName = remember(name, isLocal) { modelPickerDisplayName(name, isLocal) }
-    val provider = when {
-        modelId.startsWith(com.echoflow.data.CustomProviderConfig.PREFIX_OPENAI) -> "Direct OpenAI API"
-        modelId.startsWith(com.echoflow.data.CustomProviderConfig.PREFIX_CLAUDE) -> "Direct Claude API"
-        modelId.startsWith(com.echoflow.data.CustomProviderConfig.PREFIX_GEMINI) -> "Direct Gemini API"
-        modelId.startsWith(com.echoflow.data.CustomProviderConfig.PREFIX_CEREBRAS) -> "Direct Cerebras API"
-        modelId.startsWith(com.echoflow.data.CustomProviderConfig.PREFIX_OLLAMA) -> "Ollama API"
-        modelId.startsWith(com.echoflow.data.CustomProviderConfig.PREFIX_OPENAI_COMPATIBLE) -> "OpenAI-compatible API"
-        isLocal -> "Runs on this device — private & offline"
-        modelId.contains("/") -> modelId.substringBefore("/").replaceFirstChar { it.uppercase() }
-        else -> "Custom"
-    }
-    Surface(
-        onClick = onClick,
-        shape = MaterialTheme.shapes.large,
-        color = if (selected) MaterialTheme.colorScheme.primaryContainer else MaterialTheme.colorScheme.surfaceContainerHigh,
-        modifier = Modifier.fillMaxWidth(),
-    ) {
-        Row(Modifier.padding(Spacing.base), verticalAlignment = Alignment.CenterVertically) {
-            BrandMark(size = 40.dp)
-            Spacer(Modifier.width(Spacing.base))
-            Column(Modifier.weight(1f)) {
-                Row(verticalAlignment = Alignment.CenterVertically) {
-                    Text(
-                        displayName,
-                        style = MaterialTheme.typography.titleSmall.copy(fontWeight = FontWeight.Bold),
-                        color = if (selected) MaterialTheme.colorScheme.onPrimaryContainer else MaterialTheme.colorScheme.onSurface,
-                        maxLines = 1,
-                        overflow = TextOverflow.Ellipsis,
-                        modifier = Modifier.weight(1f, fill = false),
-                    )
-                    if (isLocal) {
-                        Spacer(Modifier.width(Spacing.s))
-                        Surface(shape = CircleShape, color = MaterialTheme.colorScheme.tertiaryContainer) {
-                            Row(
-                                Modifier.padding(horizontal = Spacing.s, vertical = 2.dp),
-                                verticalAlignment = Alignment.CenterVertically,
-                            ) {
-                                Icon(Icons.Default.OfflineBolt, null, Modifier.size(12.dp), tint = MaterialTheme.colorScheme.onTertiaryContainer)
-                                Spacer(Modifier.width(3.dp))
-                                Text("Local", style = MaterialTheme.typography.labelSmall, color = MaterialTheme.colorScheme.onTertiaryContainer)
-                            }
-                        }
-                    }
-                }
-                Text(
-                    provider,
-                    style = MaterialTheme.typography.bodySmall,
-                    color = if (selected) MaterialTheme.colorScheme.onPrimaryContainer else MaterialTheme.colorScheme.onSurfaceVariant,
-                    maxLines = 1,
-                    overflow = TextOverflow.Ellipsis,
-                )
-            }
-            if (selected) Icon(Icons.Default.CheckCircle, null, tint = MaterialTheme.colorScheme.onPrimaryContainer)
-        }
-    }
-}
-
-private fun modelPickerDisplayName(name: String, isLocal: Boolean): String {
-    if (!isLocal || name.length <= 24) return name
-
-    var clean = name
-        .removeSuffix(".gguf")
-        .removeSuffix(".GGUF")
-        .removeSuffix(".task")
-        .removeSuffix(".TASK")
-        .removeSuffix(".litertlm")
-        .removeSuffix(".LITERTLM")
-        .trim()
-
-    val quantOrPrecisionSuffix = Regex(
-        pattern = """(?i)(?:[-_](?:q[2-8](?:[-_][a-z0-9]+){0,5}|iq[1-4](?:[-_][a-z0-9]+){0,5}|mixed[-_]?int[48]|int[48]|f16|fp16|bf16))+$"""
-    )
-    clean = clean.replace(quantOrPrecisionSuffix, "")
-
-    if (clean.length > 24) {
-        clean = clean.replace(Regex("""(?i)[-_](?:20\d{2}|2[0-9]{3}|[0-9]{4})(?=$|[-_])"""), "")
-    }
-    if (clean.length > 30) {
-        clean = clean
-            .replace(Regex("""(?i)[-_](?:gguf|k[-_]?m|instruct[-_]?gguf)$"""), "")
-            .trim('-', '_', ' ')
-    }
-
-    return clean.ifBlank { name }.takeIf { it.length >= 8 } ?: name
 }
